RIP-630, /create_site: preserve state of select-sections during workflow
